+++ crates/mimic/src/design/base/types/ident.rs
@@ -1,0 +1,59 @@
+use crate::design::{base::validator, prelude::*};
+
+///
+/// Constant
+///
+
+#[newtype(
+    primitive = "Text",
+    item(prim = "Text"),
+    ty(
+        validator(path = "validator::text::len::Range", args(1, 40)),
+        validator(path = "validator::text::case::UpperSnake"),
+    )
+)]
+pub struct Constant {}
+
+///
+/// Field
+///
+
+#[newtype(
+    primitive = "Text",
+    item(prim = "Text"),
+    ty(
+        validator(path = "validator::text::len::Range", args(2, 40)),
+        validator(path = "validator::text::case::Snake"),
+    )
+)]
+pub struct Field {}
+
+///
+/// Function
+///
+/// 30 characters, snake_case
+///
+
+#[newtype(
+    primitive = "Text",
+    item(prim = "Text"),
+    ty(
+        validator(path = "validator::text::len::Range", args(2, 64)),
+        validator(path = "validator::text::case::Snake"),
+    )
+)]
+pub struct Function {}
+
+///
+/// Variant
+///
+
+#[newtype(
+    primitive = "Text",
+    item(prim = "Text"),
+    ty(
+        validator(path = "validator::text::len::Range", args(1, 40)),
+        validator(path = "validator::text::case::UpperCamel"),
+    )
+)]
+pub struct Variant {}